refactor: move the routing class from plugins to the routes package
Move as its not a plugin, instead its the top level provider of all the routes, so move it to where
it makes sense.
