+++ src/main/kotlin/com/chesire/orochi/routes/Routing.kt
@@ -1,12 +1,11 @@
-package com.chesire.orochi.plugins.routing
+package com.chesire.orochi.routes
 
 import com.chesire.orochi.routes.kitsu.kitsu
 import com.chesire.orochi.routes.status.status
 import io.ktor.application.Application
-import io.ktor.client.HttpClient
 import io.ktor.routing.routing
 
-fun Application.configureRouting() {
+fun Application.startRouting() {
     routing {
         status()
         kitsu()
